requirements.txt
backports.functools-lru-cache==1.6.4
blinker==1.6.2
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
filelock==3.12.0
Flask==2.3.2
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
indic-transliteration==2.3.44
itsdangerous==2.1.2
Jinja2==3.1.2
lit==16.0.5
llvmlite==0.40.0
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
regex==2023.5.5
requests==2.31.0
rfc3986==1.5.0
roman==4.1
sniffio==1.3.0
sympy==1.12
tiktoken==0.3.3
toml==0.10.2
torch==2.0.1
tqdm==4.65.0
triton==2.0.0
typer==0.9.0
typing_extensions==4.6.2
urllib3==2.0.2
Werkzeug==2.3.4